Separate metered-db stuff to break dependency cycle, and fix a few compilation / unit test problems
