+++ db/common/pom.xml
@@ -39,10 +39,6 @@
         </dependency>
         <dependency>
             <groupId>org.commonjava.indy</groupId>
-            <artifactId>indy-subsys-metrics-reporter</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.commonjava.indy</groupId>
             <artifactId>indy-subsys-flatfile</artifactId>
         </dependency>
     </dependencies>
