+++ pom.xml
@@ -212,6 +212,11 @@
       </dependency>
       <dependency>
         <groupId>org.commonjava.indy</groupId>
+        <artifactId>indy-db-metrics</artifactId>
+        <version>1.9.7-SNAPSHOT</version>
+      </dependency>
+      <dependency>
+        <groupId>org.commonjava.indy</groupId>
         <artifactId>indy-db-common</artifactId>
         <version>1.9.7-SNAPSHOT</version>
       </dependency>
